requirements.txt
openai
pydantic
pandas
beautifulsoup4
python-dotenv
pypdf2
